LayoutTests/fast/events/drag-and-drop-set-drag-data-arguments-expected.txt

Tests correct behavior of event.dataTransfer.setData with wrong number of arguments

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.dataTransfer.setData() threw exception TypeError: Not enough arguments.
PASS event.dataTransfer.setData("Text") threw exception TypeError: Not enough arguments.
PASS successfullyParsed is true

TEST COMPLETE
Drag me
